recipes-graphics: wayland: weston-init.bbappend: fix weston-terminal icons disappearing
When there are no [launcher] entries active in the weston.ini file,
weston will automatically add default launchers, one of which is weston-terminal.

However, if even a single [launcher] is activated, weston disables this fallback option.
Therefore, we need to re-add weston-terminal to the weston.ini file.

terminal.png is from below path of 22.04 ubuntu
/usr/share/icons/Yaru/32x32/apps/terminal.png